refactor(installer): rename bin/gaia-ui.mjs → .cjs, reuse shared module
Per desktop installer plan §3.1, the npm CLI entry point and the
Electron main process both need to share install logic, but main.cjs
is CommonJS and .mjs is ESM. Rename the bin script to .cjs so it can
require() the shared backend-installer module directly, eliminating
duplicated install logic between the two entry points.

- bin/gaia-ui.cjs now delegates all install logic to
  services/backend-installer.cjs (ensureBackend with onProgress
  streaming to stdout)
- All logs go to ~/.gaia/electron-install.log — same file the Electron
  app writes, so manual/npm installs share a unified log
- package.json bin map updated: "gaia-ui": "bin/gaia-ui.cjs"
- The file extension change is invisible to users since the bin entry
  is the only public reference to the script
